// src/app/interfaces/user.ts
/** definition class and atributes */
export interface User{
 id_user?: number; 
 create_uid?:number;  
 create_date?: Date;
 write_date?: Date; 
 name_usr: string;
 lastname_usr:string;  
mail_usr: string;  
 street_usr: string ;
 tef_usr: string;  
 password_usr?: string; 
 state_usr?: string 

}